% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_miss.R
\name{count_miss}
\alias{count_miss}
\title{Number of missing values in data}
\usage{
count_miss(data)
}
\arguments{
\item{data}{}
}
\value{
An integer number of total missing values in data
}
\description{
Number of missing values in data
}
\examples{

}
