% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_mlr.R
\name{estimate_mlr}
\alias{estimate_mlr}
\title{Estimate missing values using multiple linear regressions}
\usage{
estimate_mlr(data, diff = 1e-05)
}
\arguments{
\item{data}{A matrix or data frame object. It should have variables as columns and observations as rows.}

\item{diff}{A number. It is the minimum difference between the estimated value and the value of the immediately before iteration. It works as an exit test for the iterations, smaller its value, the estimation will be more acurate, however it will take a longer time to be estimated. \code{diff} is by default 1 e-05.}
}
\value{
A list object with positions of missing values (positions), estimated values by multiple linear regression (est_values) and new data with estimated values (new_data).
}
\description{
Estimate missing values using multiple linear regressions
}
