% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ridge.R
\name{estimate_ridge}
\alias{estimate_ridge}
\title{Estimate missing values using Ridge regression}
\usage{
estimate_ridge(data)
}
\arguments{
\item{data}{A matrix or data frame object. It should have variables as columns and observations as rows.}
}
\value{
A list object with positions of missing values (positions), estimated values by multiple linear regression with Ridge penalization (est_values) and new data with estimated values (new_data).
}
\description{
Estimate missing values using Ridge regression
}
