% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailedTree.R
\name{example_tree}
\alias{example_tree}
\title{example_tree() function}
\usage{
example_tree()
}
\description{
Creates a new object of the "detailedTree" class using an example tree with 9 taxa
}
\details{
The example tree has 9 taxa and a height of 14. It is very unbalanced and should only be used for testing purposes.
}
