% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_mean.R
\name{impute_mean}
\alias{impute_mean}
\title{Simple mean imputation of missing values}
\usage{
impute_mean(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Simple mean imputation of missing values
}
