% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_median.R
\name{impute_median}
\alias{impute_median}
\title{Simple median imputation of missing values}
\usage{
impute_median(data)
}
\arguments{
\item{data}{A matrix or data frame object. It should have variables as columns and observations as rows.}
}
\value{
A list object with positions of missing values (positions), imputed values by observed median (imp_values) and new data with imputed values (new_data).
}
\description{
Simple median imputation of missing values
}
