% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{make_MCMC}
\alias{make_MCMC}
\title{make_MCMC function}
\usage{
make_MCMC(
  dataTree = NULL,
  filepath = ".",
  prior_sigma_x = distributions3::Uniform(0, 1),
  prior_sigma_y = distributions3::Uniform(0, 1),
  prior_sigma_r = distributions3::Uniform(0, 1),
  prior_sigma_s = distributions3::Uniform(0, 1),
  prior_sigma_a = distributions3::Uniform(0, 5),
  prior_root_x = distributions3::Normal(0, 1),
  prior_root_y = distributions3::Normal(0, 1),
  prior_root_r = distributions3::Normal(0, 5),
  prior_root_s = distributions3::Normal(0, 5),
  prior_root_a = distributions3::Normal(0, 5),
  prior_mu = distributions3::Normal(0, 5),
  prior_kappa = distributions3::Uniform(0, 5),
  proposal_weights = list(sigma_x = 10, sigma_y = 10, sigma_r = 10, sigma_s = 10, sigma_a
    = 10, root_x = 1, root_y = 1, root_r = 1, root_s = 1, root_a = 1, mu = 10, kappa =
    10, W_d = 10, W_m = 10, W_c = 10, W_h = 10, V_r = 20, V_s = 20, V_a = 20, tip = 10),
  under_prior = FALSE
)
}
\arguments{
\item{dataTree}{An object of class "dataTree"}

\item{filepath}{An object of class "character" (a string) that gives the filepath where you want the output log files saved. Default behavior will save the output to the same directory where the mcmc is run.}

\item{prior_sigma_x}{Prior on the sigma_x model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Uniform(0,1)}

\item{prior_sigma_y}{Prior on the sigma_y model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Uniform(0,1)}

\item{prior_sigma_r}{Prior on the sigma_r model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Uniform(0,1)}

\item{prior_sigma_s}{Prior on the sigma_s model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Uniform(0,1)}

\item{prior_sigma_a}{Prior on the sigma_a model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Uniform(0,5)}

\item{prior_root_x}{Prior on the root_x model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Normal(0,1)}

\item{prior_root_y}{Prior on the root_y model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Normal(0,1)}

\item{prior_root_r}{Prior on the root_r model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Normal(0,5)}

\item{prior_root_s}{Prior on the root_s model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Normal(0,5)}

\item{prior_root_a}{Prior on the root_a model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Normal(0,5)}

\item{prior_mu}{Prior on the mu model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Normal(0,5)}

\item{prior_kappa}{Prior on the kappa model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Uniform(0,5)}

\item{proposal_weights}{A list of MCMC proposals and their relative weights (can use proportions or whole numbers). If you want to exclude a proposal, the weight can be set to 0}

\item{under_prior}{Whether to perform MCMC under the prior (acceptance ratio independent of likelihood), default is FALSE}
}
\description{
Creates a new instance of the ellipseMCMC class
}
