% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numb_miss.R
\name{numb_miss}
\alias{numb_miss}
\title{Title Number of missing values in data}
\usage{
numb_miss(data)
}
\arguments{
\item{data}{}
}
\value{
An integer number of total missing values in data
}
\description{
Title Number of missing values in data
}
