% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_miss.R
\name{pos_miss}
\alias{pos_miss}
\title{Missing value position inside data set}
\usage{
pos_miss(data)
}
\arguments{
\item{data}{A matrix or data frame object. It should have variables as columns and observations as rows.}
}
\value{
A data frame with the row (Row) and column (Col) position of each missing value.
}
\description{
Missing value position inside data set
}
