% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTree.R
\name{save_dataTree}
\alias{save_dataTree}
\title{save_dataTree function}
\usage{
save_dataTree(tree, filepath = ".", prefix = "sim")
}
\arguments{
\item{tree}{An object of class "dataTree" to be saved}

\item{filepath}{An object of class "character" giving the filepath to the directory where you want the output to be saved. The default behavior will save files to the current directory.}

\item{prefix}{An object of class "character" giving the prefix you want to assign to output files. The default prefix is "sim".}
}
\description{
This functions saves each element of a dataTree in a speficied directory with a specified prefix
}
\details{
The output files are: <prefix>.tree.txt, <prefix>.W.tsv, <prefix>.V.tsv, <prefix>.XY.tsv, and <prefix>.param.tsv
}
