% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_distr.R
\name{sim_distr}
\alias{sim_distr}
\title{Simulate data variables with known distributions}
\usage{
sim_distr(n, k = 3, distr, ...)
}
\arguments{
\item{n}{number of observations}

\item{k}{number of variables}

\item{...}{}
}
\value{
sim_distr gives a data frame with the observations as rows and the variables as columns
}
\description{
Simulate data variables with known distributions
}
