% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_mlr.R
\name{estimate_mlr}
\alias{estimate_mlr}
\title{Estimate missing values using multiple linear regressions}
\usage{
estimate_mlr(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Estimate missing values using multiple linear regressions
}
