% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_mlr.R
\name{estimate_mlr}
\alias{estimate_mlr}
\title{Estimate missing values using multiple linear regressions}
\usage{
estimate_mlr(data)
}
\arguments{
\item{data}{A matrix or data frame object. It should have variables as columns and observations as rows.}
}
\value{
A list object with positions of missing values (positions), estimated values by multiple linear regression (est_values) and new data with estimated values (new_data).
}
\description{
Estimate missing values using multiple linear regressions
}
