% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ridge.R
\name{estimate_ridge}
\alias{estimate_ridge}
\title{Estimate missing values using Ridge regression}
\usage{
estimate_ridge(data, diff = 1e-05, ridge_alpha = 0)
}
\arguments{
\item{data}{A matrix or data frame object. It should have variables as columns and observations as rows.}

\item{diff}{A number. It is the minimum difference between the estimated value and the value of the immediately before iteration. It works as an exit test for the iterations, smaller its value, the estimation will be more accurate, however it will take a longer time to be estimated. \code{diff} is by default 1 e-05.}

\item{ridge_alpha}{It is a specific alpha value applied in the cross validation of Ridge regression. It could give a different penalization and therefore a different estimated value. alpha = 0 by default.}
}
\value{
A list object with positions of missing values (positions), estimated values by multiple linear regression with Ridge penalization (est_values) and new data with estimated values (new_data).
}
\description{
Estimate missing values using Ridge regression
}
