% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_median.R
\name{imp_median}
\alias{imp_median}
\title{Simple mean imputation of missing values}
\usage{
imp_median(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Simple mean imputation of missing values
}
