% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_median.R
\name{impute_median}
\alias{impute_median}
\title{Simple median imputation of missing values}
\usage{
impute_median(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Simple median imputation of missing values
}
