% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_distr.R
\name{sim_distr}
\alias{sim_distr}
\title{Simulate data variables with known distributions}
\usage{
sim_distr(n, k, distr, ...)
}
\arguments{
\item{n}{Number of observations.}

\item{k}{Number of variables.}

\item{distr}{A vector of distributions to be used in each variable. It could be "normal", "exponential", "poisson", "weibull", "chisquare" or "uniform". In case to use the same distribution for all variables, set it as distr = "distr".}

\item{...}{Arguments specific to each distribution. See \code{\link[stats:rnorm]{rnorm}}, \code{\link[stats:rexp]{rexp}}, \code{\link[stats:rpois]{rpois}}, \code{\link[stats:rweibull]{rweibull}}, \code{\link[stats:rchisq]{rchisq}}, \code{\link[stats:runif]{runif}} to know these arguments.}
}
\value{
\code{sim_distr} gives a data frame with the observations as rows and the variables as columns
}
\description{
Simulate data variables with known distributions
}
\seealso{
\code{\link[stats:rnorm]{rnorm}},
\code{\link[stats:rexp]{rexp}},
\code{\link[stats:rpois]{rpois}},
\code{\link[stats:rweibull]{rweibull}},
\code{\link[stats:rchisq]{rchisq}},
\code{\link[stats:runif]{runif}}.
}
