% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_norm.R
\name{sim_norm}
\alias{sim_norm}
\title{Simulate data with or without normality
Generates simulated data sets with normality (rnorm) or without normality (rpois).}
\usage{
sim_norm(
  n,
  k = 3,
  mean.v = rep(0, k),
  sd.v = rep(1, k),
  lambda,
  normality = TRUE
)
}
\arguments{
\item{n}{Number of individuals}

\item{k}{Number of variables}

\item{mean.v}{A vector of means}

\item{sd.v}{A vector of standard deviations}

\item{lambda}{Lambda value for Poisson distribution}

\item{normality}{Normality in data set}
}
\value{
data.frame
}
\description{
Simulate data with or without normality
Generates simulated data sets with normality (rnorm) or without normality (rpois).
}
\examples{
#Normal data set
data_ej_norm <- sim_norm(n = 100,k = 3, normality=TRUE)
sapply(data_ej, shapiro.test)
#No normal data set
data_ej_no_norm <- sim_norm(n = 100, lambda = 4, normality = FALSE)
sapply(data_ej_no_norm, shapiro.test)
}
